Restructure Button styled-component
